Add `CustomDump` trait
Enabled by default for backwards-compatibility, this trait allows folks
to opt out of the library.
